refactor(table): movendo goEdit para exemplo
