+++ src/components/Table/Example.jsx
@@ -1,16 +1,25 @@
 import Table from "./Table";
 import GenericContextProvider from "../../contexts/GenericContext";
+import { useNavigate } from "react-router-dom";
 
 function Example() {
+  const navigate = useNavigate();
+  const routeName = "contrato";
+
+  const goEdit = (id) => {
+    navigate(`/${routeName}/${id}/editar`);
+  };
+
   return (
     <GenericContextProvider lSName="contratos">
       <Table
-        routeName="contrato"
+        routeName={routeName}
         headerCells={[
           "Nome",
           "Valor Total",
           "Tipo de Vencimento",
           "Vencimento",
+          "",
         ]}
         registerLabel="Criar Contrato"
       >
@@ -28,6 +37,18 @@ function Example() {
                 : element.typeExpire === "por Tempo"
                 ? " meses"
                 : ""}
+            </td>
+            <td>
+              <button
+                className="btn-icon-edit"
+                onClick={(e) => {
+                  e.stopPropagation();
+                  goEdit(element.id);
+                }}
+                title="Editar"
+              >
+                <i className="bi bi-pencil-fill"></i>
+              </button>
             </td>
           </>
         )}
